kernel/random: Replace kernel RNG with sponge function
We can now mix arbitrary amounts of entropy and generate an
arbitrary amount of randomness.

+++ kernel/shellutils/rand.cpp
@@ -3,9 +3,9 @@
 #include <libk/printf.h>
 
 void shell_rand(char *) {
-  auto num = Kernel::random_prng<int16_t>();
+  auto num = Kernel::random_prng<uint32_t>();
 
   Kernel::Drivers::VGATerminal::lock();
-  kprintf("Your random number is %d\n", num);
+  kprintf("Your random number is %u\n", num);
   Kernel::Drivers::VGATerminal::unlock();
 }
